Update IntegrationTestCase tests to respect new logic.
In `::testArrayUrls()`, we declare an explicit `_method` to bypass the early GET-only route declared in `::setUp()` that would otherwise now match first.

In `::testAssertRedirect()`, we update the Location in Response we're testing against (since it's a GET), and then update the static string URL to match. We do this since we want the array-based reverse route lookup to match the earlier `/get/tasks/index` route now.