Fix route reset and preserve extensions
I wasn't entirely happy with the solution in #12695 and wanted to try an
alternate approach. This set of changes adds state to fewer places and
doesn't require plugins to opt-in to backwards compatible behavior.
Instead it introduces new internal methods to reset router state that
can cause conflicts/duplicate work. Global extensions and default route
classes are not reset.

Refs #12695
Refs #12577