Try to fix codecov builds
Don't include TRAVIS_PHP_VERSION in the commands that run coverage. Due
to codecov/codecov-bash#133 we should try to not have a version number
in the command so that codecov can detect environment variables better.

This will add more builds, but we might get coverage for all versions of
PHP & all database drivers.