Set "autoSetCustomValidity" to false by default.
That way it works it's likely to be undesirable by most users.
For e.g. if you have a required email field, by default the browser
will show message like "Please fill out this field" if you leave
the field blank.  If you enter an invalid email string it will show
"Please enter an email".

With the custom validity message feature it always shows same message
"This field cannot be left empty". It takes no account of additional
validations for the field even if set.