Fix optional routing keys with braced placeholders
Optional parameters were not working correctly with braced parameters.
While non-core parameters can be optional the controller + action
parameters should never be null as that will result in invalid URLs
generated or incorrect results that only contain the action parameter.

I've removed a few invalid test cases that were providing keys as both
in the template and as a default. This has not worked for a long time
and these tests were relying on the now invalid behavior.

Port changes from #14089 to 4.x
Fixes #13997