Fix trace/context for errors wrapped by debug kit
Add a magic `_trace_offset` context value that allows error handling
wrappers to offset the stack trace that is displayed. This is not
a great solution, but given how PHP error handling works and our
backwards compatibility requirements this is what I could come up with.

Port changes from #14090 to 4.x