Start Html Debug output
These changes add an unstyled HTML dump formatter and does some cleanup
in the console/text formatters. I will add CSS and Javascript to the
HTML dump in an upcoming pull request as to not make this one really
big.

I have marked all the output formatters and their interface as
internal/unstable. I would like to introduce the concept of unstable
APIs into CakePHP as a way to help us be able to ship features without
being required to fully commit to not changing them. I envision unstable
to be something that lasts on an class/interface for one minor release.