Fix withStatus() overwriting content-type
The withStatus() method would use the `_contentType` property which can
contain stale information if `withHeader()` is used to set the
content-type. Instead of storing potentially divergent information we
can make the content-type header authoritative and do type munging in
getType().

While this could be a breaking change we make precious few promises
around protected properties. This change should be backported to 3.x as
well to resolve the same problem there.

Backport fixes from #14211 to 3.x