Fix null values in Hash::get
Did not work as described. The default value should be returned when the path does not exist according to the docs. Because of the use of isset it will also return the default value when the provided field holds null as value. Since already checked if its an array its better to use array_key_exists.