Improve ErrorLogger for PHP errors
I was integrating an error tracking service with a CakePHP application
and had hoped to be able to replace the `errorLogger` with a different
implementation that forwarded errors to the remote service and then also
logged to syslog.

This was challenging because exceptions are logged by the ErrorLogger
but PHP errors are logged via the ErrorHandler. This change unifies
logging to always be done in the ErrorLogger and adds an interface so we
don't have to be lucky when implementing custom error loggers.