Add session based CSRF token middleware
The current double submit cookie CSRF middleware has a few potential
weaknesses:

1. The tokens it uses are portable between users. If a user has their
   browser compromised, CSRF tokens can be re-used between other
   compromised users.
2. Tokens don't have an expiration time. If a CSRF token is leaked and
   a user is compromised an old token can be used again.

These changes introduce an alternative CSRF middleware that stores the
compare token in the session. This removes the need to do HMAC signing
as the compare value is in the session where an attacker cannot change
it. Furthermore, it also couples CSRF tokens to the user, and to the
specific session. This mitigates both weaknesses with our existing CSRF
middleware.

I didn't want to make the existing CSRF middleware more complicated and
risk breaking backwards compatibility which is why this is a new
middleware. A separate middleware makes opt-ing into session based
tokens explicit.