Fix output buffers not being closed when the view triggers an exception.
When evaluating a template or caching a block, the view itself opens a
buffer, and the templates/blocks that it evaluates might open buffers
too. In case an exception happens, the opened buffers will not be
closed, causing risky tests with PHPUnit, and possible further output to
be swallowed unless the test itself closes/flushes the remaining
buffers.