Allow TimeType to marshal H:i format
This aligns the behavior of TimeType with the values accepted by
Validation::time(). Doing so will prevent tedious to diagnose problems
where validation passes but no value is set in the entity.

The x:y format could also be marshaled as I:S but we already accept
Y-m-d H:i in the DateTimeType and having consistency is important.

Fixes #14957