Salt stateless CSRF tokens as well
Add salting to CSRF tokens to prevent BREACH weakness. This ports the
changes made to session based CSRF tokens to stateless ones as well.
This will result in longer tokens as both the cookie and the POST data
are base64 encoded, and the POST data is now 112 bytes instead of 56
bytes.