Fix CSRF token backwards compatibility
I did more thorough testing with salted CSRF tokens and this change is
required to provide backwards compatibility. Because the old tokens are
hexadecimal, they also decode as base64 data. When an old token is found
in the cookie we shouldn't base64 decode or apply salting, as the data
becomes hard to manage in subsequent requests.