Add a schema loader for CakePHP core tests
I was planning on using phinx and a single phinx migration for this,
however that is not going to work. Phinx depends on cakephp/database
which cannot be installed as it would conflict with the root package.

Because of that limitation I've added a way to use the Database
package's schema generation features more directly. Retaining this long
term will term will require having some schema generation capabilities
in CakePHP. Those features could be extracted out into an internal package
for 5.x.