Drop constraints during reset and reflect schema again
SQLServer cannot drop tables if they are referenced in constraints.
Before trying to drop tables we need to remove constraints. Furthermore,
SQLServer needs to have the full schema reflection in order to reset
sequences when truncating. This schema can be cached though as we don't
want to reflect schema on each test.