Move fixture reset management to test case
Having the reset strategy be global across the application complicates
a few scenarios:

1. If an application has some end-to-end tests they would be forced to
   use `TruncationStrategy` even for their integration tests.
2. Having multiple state managers that are used in different tests gets
   harder. With this approach if a test only uses elasticsearch it can
   use the elasticsearch state reset (once it exists).

In addition to these specfic problems having the state manager on the
test case makes more of the framework's logic extensible and replaceable
should an application need to.

For the core tests having this will unblock us to use the
`TransactionStrategy` on a large number of our tests which should yield
even faster test durations.