Add state reset aggregate class
Add an aggregate/collection class for state reset objects. This will be
necessary for applications that use multiple fixture providers.

An example of this would be an application that has tests that insert data
into both a relational database and elasticsearch. The application could
define the `getStateResetStrategy()` method and return the aggregate
object with a database and elastic search object inside.