Move fixture cleanup to teardown
Doing cleanup in setup causes problems for subsequent test runs as data
from the previous test run is left behind in the database. Doing the
cleanup in teardown attempts to ensure that the database is cleared at
the end of the test.

This will work less well when tests are killed before a test run is
complete as state will not be cleaned up. The old fixture system solved
this by dropping tables at the beginning of a test run which can happen
in the new fixture system depending on how schema is loaded.