Fix problems with aria-describedby and remove aria-label
The aria-describedby attribute can only be set properly when we know
that the error text will display and that the error message will have an
id attribute set.

I've also removed the aria-label attribute. Our forms use labels and the
`for` attribute which is enough for assitive technologies to provide
useful feedback.