Remove redundant listTablesAndViews method
I feel that this method can be removed and `listTables()` can be
un-deprecated if we normalize all of the driver dialects. The only
divergent implementation was sqlite. This driver is rarely used for
production scenarios but is a great in local development and CI.
Because of how SQLite is used I think the risk is low and its divergence
is not documented and can be interpreted as a defect. I recognize that
we generally don't update the code to the docs but I think this
situation is more complex. There exist 3 other implementations of the
same method that all work differently and all 4 are documented the same
way makes me believe that we can treat this as a bug.

I also would like to add as little API scope as possible as we are
shipping this method in a bug fix. The reason for that is that I see
views being included in truncation loops as a defect in the new fixture
system and the only solution is to add a new method :shrug: