Add tests for ExceptionTrap
Add a TextRenderer for making it possible to test Exception trap in CI.

The interface for ExceptionRenderInterface needs to change, but we
can't do that without breaking changes. Instead use method_exists()
checks to give weaker emulation of the future interface.

I don't love this approach but adding another interface seemed heavy as
well because the good name is already in use.