4.4 - ExceptionTrap::instance() (#16278)
Add global state to ExceptionTrap

I don't really like global state, but in this situation we have natural
global state that interfaces with PHP. Having an `instance()` method
will allow middleware to get the current globally registered instance.

This may seem a bit dumb as we could make another ExceptionTrap in the
middleware. My concern with this that developers would need to add their
callbacks twice if they were using that feature which is also new.

There is an argument to remove callback support as userland code could
build a renderer/logger that handle their needs.

Co-authored-by: othercorey <corey.taylor.fl@gmail.com>