Add deprecation for RequestHandlerComponent
This follows the approach we took with `AuthComponent`. RequestHandler
is used in *many* applications and the new features are still pretty
fresh. A soft deprecation will give users more time to upgrade as the
upgrade requires non-trivial changes.

The docs mention a `CheckHttpCacheComponent`. That component still needs
to be written though.