Add deprecation to BaseErrorHandler::register()
This may be too aggressive of a deprecation. However, the solution is to
cut 7 lines from the application bootstrap and replace it with 4 lines.

This upgrade will allow DebugKit to deliver an improved
deprecation panel, and makes building custom error handling easier as
the renderers provide a simpler interface to implement. The upgrade
also enables error rendering, logging and configuration for the entire
subsystem from configuration files with no need to inherit code from
the framework, hopefully giving more future proof APIs that will break
less frequently.