4.4 - Move exception renderer selection to a renderer
Add a new exception renderer that handles selecting the console/html
renderers based on the current SAPI. This lets us remove that logic from
`ExceptionTrap`, and gives users a simple interface based hook to
implement their own renderer selection if they want to do that. I'm not
in love with the name but it reflects what it does.

I've deprecated and renamed `ExceptionRenderer` to
`HtmlExceptionRenderer` as that is closer to what it actually does.
This solution will still require userland changes to use. I attempted to
make `ExceptionRenderer` the 'selector' class but that would have
resulted in significant compatibility breaks as all of the internals
would have been removed. The deprecation based approach is
disruptive to userland code but they have a simpler class rename change
to make.

If folks like this approach I can replicate it to `ErrorTrap` as well.