Make ExceptionTrap handle old configuration better.
The attempt in #16445 didn't work so well. This is another attempt that
retains the API compatibility with `ExceptionRenderer` by making it
a sub-class of the new name `WebExceptionRenderer`. Furthermore, the
code now assumes that if you are using the old name that your
configuration is 'wrong' and that in console environments you actually
want to use the console exception renderer.

No changes are made for configurations that explicitly define an
`exceptionRenderer` key. These setups could still emit HTML in console
environments if users do not also upgrade their configuration.