Add fallbacks for `viewClasses` and a view to require negotiation.
The `NegotiationRequiredView` allows an application that requires
content-type negotiation to avoid the default HTML error page rendering
and emit a smaller (and faster) response. While not exceptionally useful
it is a good baseline implementation for the more important feature of
having a user-land way to define a fallback view class.

Fixes #16467