Further improve the ErrorLoggerInterface
Provide better method names, and better parameters. This should also
help reduce the need for error handling configuration to be used in
Loggers which feels like a win as userland code doesn't need to read
configuration data to make decisions on whether or not to include
a trace.

The new methods can't be added to the interface because of backwards
compatibility so I've annotated them and used `method_exists()`. This
isn't ideal but it does allow us to be backwards compatible.