4.5 Add deprecations for most of ModelAwareTrait
Adding deprecations to loadModel() is tricky. While it is important for
this method to emit a deprecation so that developer can fix their usage,
we also use it internally quite extensively to load ORM models.

We might need a more elaborate solution to emit deprecations for this
method. One option that comes to mind is only emitting deprecations for
non-ORM usage. We *could* have a `loadModel()` method in 5.0 that emits
deprecations.