Add fetchModel() and un-deprecate loadModel()
When figuring out how we could add deprecations to loadModel() we
uncovered that this would be very hard because of how loadModel() is
used. The deprecation of loadModel() was primarily driven by wanting to
avoid the 8.2 dynamic property deprecation from PHP. We can add
a deprecation ourselves for that problem with a suggestion on how to
resolve this problem.

Deprecating loadModel() would also create a rough upgrade path to other
locator based implementations (like in usemuffin/webservices). Instead
we can retain loadModel() and add fetchModel(). Throughout 5.x we can
encourage users stop using loadModel(), and start using other methods.