Improve rendering of error pages
PHP exceptions don't include the source line/file in the stacktrace so
our error pages started on the frame *after* the line where the error
actually was. This makes debugging harder as the problematic line isn't
as obvious.

These changes add an unclosable frame before each stack that shows the
line the exception came from.