Update query logging for PHPUnit changes
Because PHPUnit no longer has a `--debug` option, we need to change
how query logging works in test runs. Instead of reading the PHPUnit
flag, we read an environment variable. This can either be set in the
phpunit.xml file or in the execution environment.

Once we're happy with this, I'll update the app skeleton, and migration
guide.