Reverse connection aliasing logic to restore previous behavior
In #17261 we reversed the connection aliasing behavior so that if you
had both a test and non test connection defined, and used the `test_`
prefixed names in fixtures your local development database would be
truncated.

The logic historically was that the framework would require you to
define test_ prefixed connections and aliases would be made from live ->
test so that accessing the live connection would retrieve the test
connections. Having test prefixed connection names resolve to the real
connection makes it hard to maintain a persistent development
environment.

Fixes #17619