Fix enableCsrfToken causing GET requests to fail
If CSRF tokens were enabled, and a POST, then a GET, the GET request
would fail. By passing the method forward to `_addTokens` we can make
better decisions on when to add the CSRF token to the request data. CSRF
tokens do not need to be present in the request data of OPTIONS or GET
requests.

Fixes #17612