Being adding SchemaDialect::columnDefinitionSql
In order to remove duplication between migrations and cakephp/database,
we need a method to generate column SQL fragments for DDL operations.
The current method in SchemaDialect isn't well suited for this as it
requires an `TableSchema` instance which we don't always have in
migrations.

This new with consumes the output of describeColumns() which migrations
will also start using. I've also added `onUpdate` to `TableSchema`
reflection for datetime/timestamp columns so that schema generation is
consistent between the two APIs. These changes only contain an
implementation for MySQL. Other drivers will be implemented in follow up
pull requests.