Throw exception when using plugin prefixed names.
ObjectRegistry does not store objects with plugin prefixed names.
So accepting prefix names for it's load() and set() will lead to incorrect results.
Throwing an explicit exception is better than try to use the name
after silently dropping the prefix as there's no guarantee that
the object loaded using the same plugin prefix.