Add first pass of Constraint and ForeignKey objects
I would like to preserve backwards compatibility with the existing
`constraints()` method so we'll need objects for Constraints. It doesn't
feel right to have the foreign key methods available to unique keys, so
I've made a Constraint to act as a base class, and `ForeignKey` contains
the specifics for foreign keys. I would also like to include CHECK
constraints in a subsequent set of changes.