Preserve `constraint` in TableSchema
If the driver returns a `constraint` key in index reflection data, use
that as the name. This is currently only in use by postgres with custom
primary key names.

This preserves behavior necessary for migrations and improves schema
reflection data from cakephp. Ideally the primary key would be available
at its name as well.